tests/wpt/tests/html/semantics/grouping-content/the-hr-element/grouping-hr.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>the hr element</title>
    <link rel="author" title="dzenana" href="mailto:dzenana.trenutak@gmail.com">
    <link rel="help" href="https://html.spec.whatwg.org/multipage/#the-hr-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        "use strict";

        // check that prototype matches spec's DOM interface
        test(function () {
            var testElement = document.createElement("hr");
            assert_equals(Object.getPrototypeOf(testElement), HTMLHRElement.prototype, "HTMLHRElement.prototype should be used for hr");
        }, "The prototype for hr is HTMLHRElement.prototype");

        // Not checking: "The hr element does not affect the document's outline."

    </script>
</head>
<body>
    <h1>Description</h1>
    <p>This test validates the hr element.</p>

    <div id="log"></div>

</body>
</html>
